# Added memory optimisations for GetLastActiveSpan