@bruno-garcia 2023-09-20
Worth noting; These microbenchmarks works on tightloop on a single thread. This is code that will run concurrently and contention is a big concern, so just as much test on a server scenario under a load test is worth going after