@jamescrosswell 2023-09-20
> It looks like the wrong synchronization mechanism (RWLock seems to be a better fit) but since this isn't mostly just reads once data is written, in which case Immutable type might be better, how about `ConcurrentStack` instead?

I thought about ConcurrentStack but I need a lock around both the `Peek` and the `Pop` operations in this method (to ensure another thread doesn't do a `Push` between calling `Peek` and `Pop`... which would cause the wrong element to be popped):

https://github.com/getsentry/sentry-dotnet/blob/78871ba82e4758a606610a876321f945da42b5ae/src/Sentry/TransactionTracer.cs#L335-L352

So I think just a vanilla `Stack<T>` with a `lock` is the way to go.