@jamescrosswell 2023-09-21
@bruno-garcia off the back of our conversation about upgradable locks, I took a look at `ReaderWriterLockSlim.EnterUpgradeableReadLock` and I'm still not sure it will work in a multithreaded scenario. 

For example:

```mermaid
sequenceDiagram
    Thread1->>LastActiveSpanTracker: EnterUpgradeableReadLock
    Thread2->>LastActiveSpanTracker: EnterUpgradeableReadLock
    Thread1->>LastActiveSpanTracker: EnterWriteLock (Waits for reads to finish)
    Thread2->>LastActiveSpanTracker: EnterWriteLock (Also waits for reads to finish -> Deadlock!)
``` 

And I noticed this in the [in the remarks](https://learn.microsoft.com/en-us/dotnet/api/system.threading.readerwriterlockslim.enterupgradeablereadlock?view=net-7.0#remarks):

> Only one thread can enter upgradeable mode at any given time.

I assume that's to prevent the above deadlock scenario from playing out.

If I understand correctly and if we only have blocks of code leveraging either `EnterUpgradeableReadLock` or `EnterWriteLock` then I think `ReaderWriterLockSlim` is basically doing the same thing as `lock`. Have I understood this correctly or there something I'm still missing?